Allow loading plugins after seneca is ready
