+++ lib/plugins.js
@@ -1,6 +1,7 @@
 /* Copyright (c) 2014-2015 Richard Rodger, MIT License */
 'use strict'
 
+var Util = require('util')
 var _ = require('lodash')
 var Eraro = require('eraro')
 var Parambulator = require('parambulator')
@@ -38,7 +39,7 @@ module.exports = function default_decorations () {
   seneca.decorate('plugins', internals.all)
 }
 
-module.exports.register = function (callpoint) {
+module.exports.register = function (so, callpoint) {
   return function api_register (plugin) {
     var seneca = this
 
@@ -48,7 +49,7 @@ module.exports.register = function (callpoint) {
       }
     })
 
-    seneca.private$.definitions.push(function (so, next) {
+    var definition = function (so, next) {
       var fullname = plugin.name + (plugin.tag ? '/' + plugin.tag : '')
 
       plugin.fullname = fullname
@@ -124,8 +125,61 @@ module.exports.register = function (callpoint) {
         { exports: exports }, fullname !== pluginref ? fullname : undefined)
 
       next(null, plugin)
+    }
+
+    if (!seneca.private$.ready) {
+      seneca.private$.definitions.push(definition)
+      return
+    }
+
+    definition(so, function (err, plugin) {
+      if (err) {
+        throw err
+      }
+      module.exports.init(seneca, so, plugin, function (err) {
+        if (err) {
+          throw err
+        }
+      })
     })
   }
+}
+
+module.exports.init = function (seneca, so, plugin, next) {
+  seneca.act({
+    init: plugin.name,
+    tag: plugin.tag,
+    default$: {},
+    fatal$: true,
+    local$: true
+  },
+  function (err, out) {
+    if (err) {
+      var plugin_err_code = 'plugin_init'
+
+      plugin.plugin_error = err.message
+
+      if (err.code === 'action-timeout') {
+        plugin_err_code = 'plugin_init_timeout'
+        plugin.timeout = so.timeout
+      }
+
+      seneca.die(internals.error(err, plugin_err_code, plugin))
+      return next(err)
+    }
+
+    if (so.debug.print && so.debug.print.options) {
+      console.log('\nSeneca Options (' + seneca.id + '): plugin: ' + plugin.name +
+        (plugin.tag ? '$' + plugin.tag : '') + '\n' +
+        '===\n')
+      console.log(Util.inspect(plugin.plugin_options, { depth: null }))
+      console.log('')
+    }
+
+    var pluginref = plugin.name + (plugin.tag ? '/' + plugin.tag : '')
+    seneca.log.debug('register', 'ready', pluginref, out)
+    return next()
+  })
 }
 
 internals.isRegistered = function api_hasplugin (plugindesc, tag) {
